Выносит адаптер и viewholder в core
