+++ core/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="button_edit_description">редактировать</string>
+    <string name="button_delete_description">удалить</string>
+</resources>
