+++ featureStartScreen/src/main/res/layout/fragment_start_screen.xml
@@ -78,7 +78,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/my_games"
-            android:layout_marginBottom="@dimen/margin_botton"
+            android:layout_marginBottom="@dimen/margin_bottom"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent" />
