+++ featureStartScreen/src/main/res/values/strings.xml
@@ -2,7 +2,5 @@
 <resources>
     <string name="create_game">Создать игру</string>
     <string name="my_games">Мои игры</string>
-    <string name="button_edit_description">редактировать</string>
-    <string name="button_delete_description">удалить</string>
     <string name="delete_template_successfully">Шаблон успешно удален</string>
 </resources>
